Migrate most of main activity to compose
At this stage it's not possible to migrate everything as it would prevent the usage of inset APIs by the composables that are still displayed in fragments. That's because the `AndroidView` would consume the insets and not propagate them since having Compose -> View -> Compose is not ideal, so until the fragments are removed it won't be possible to migrate everything.

- Create a top app bar composable for the title and offline alert
- Create a bottom app bar layout for the nav bar and fab
- Replace the XML menus with state items
- Add support for scroll detection to mimic the coordinator layout behaviour in compose